Remove the concept of beatmap events, rename BreakEvent -> BreakPeriod.
