--- osu.Game/Beatmaps/Events/BackgroundEvent.cs
@@ -1,13 +1,0 @@
-﻿// Copyright (c) 2007-2017 ppy Pty Ltd <contact@ppy.sh>.
-// Licensed under the MIT Licence - https://raw.githubusercontent.com/ppy/osu/master/LICENCE
-
-namespace osu.Game.Beatmaps.Events
-{
-    public class BackgroundEvent : Event
-    {
-        /// <summary>
-        /// The file name.
-        /// </summary>
-        public string Filename;
-    }
-}
