--- osu.Game/Beatmaps/Events/Event.cs
@@ -1,13 +1,0 @@
-﻿// Copyright (c) 2007-2017 ppy Pty Ltd <contact@ppy.sh>.
-// Licensed under the MIT Licence - https://raw.githubusercontent.com/ppy/osu/master/LICENCE
-
-namespace osu.Game.Beatmaps.Events
-{
-    public abstract class Event
-    {
-        /// <summary>
-        /// The event start time.
-        /// </summary>
-        public double StartTime;
-    }
-}
